clean up the scripts
