+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@maxdeliso/typed-ski",
-  "version": "0.18.2",
+  "version": "0.18.3",
   "license": "MIT",
   "type": "module",
   "exports": {
